test: move HMR Error Recovery test to a separate file
